implemented feature preprocessing layer
